+++ src/app/pages/send/choose-crypto-asset/choose-crypto-asset.tsx
@@ -5,19 +5,16 @@ import { RouteUrls } from '@shared/route-urls';
 import { useRouteHeader } from '@app/common/hooks/use-route-header';
 import { useAllTransferableCryptoAssetBalances } from '@app/common/hooks/use-transferable-asset-balances.hooks';
 import { useWalletType } from '@app/common/use-wallet-type';
+import { Brc20TokensLoader } from '@app/components/brc20-tokens-loader';
 import { Header } from '@app/components/header';
-import { useBrc20TokensByAddressQuery } from '@app/query/bitcoin/ordinals/brc20/brc20-tokens.query';
-import { useCurrentAccountTaprootAddressIndexZeroPayment } from '@app/store/accounts/blockchain/bitcoin/taproot-account.hooks';
 
+import { Brc20TokenAssetList } from '../../../components/crypto-assets/bitcoin/brc20-token-asset-list/brc20-token-asset-list';
 import { ChooseCryptoAssetLayout } from './components/choose-crypto-asset.layout';
 import { CryptoAssetList } from './components/crypto-asset-list';
 
 export function ChooseCryptoAsset() {
   const navigate = useNavigate();
   const allTransferableCryptoAssetBalances = useAllTransferableCryptoAssetBalances();
-
-  const { address: bitcoinAddressTaproot } = useCurrentAccountTaprootAddressIndexZeroPayment();
-  const { data: brc20Tokens = [] } = useBrc20TokensByAddressQuery(bitcoinAddressTaproot);
 
   const { whenWallet } = useWalletType();
 
@@ -32,8 +29,15 @@ export function ChooseCryptoAsset() {
             software: true,
           })
         )}
-        brc20Tokens={brc20Tokens}
       />
+      {whenWallet({
+        software: (
+          <Brc20TokensLoader>
+            {brc20Tokens => <Brc20TokenAssetList brc20Tokens={brc20Tokens} />}
+          </Brc20TokensLoader>
+        ),
+        ledger: null,
+      })}
     </ChooseCryptoAssetLayout>
   );
 }